+++ All.ui/miniprogram/pages/index/index.js
@@ -28,7 +28,7 @@ Page({
     const { url } = currentTarget.dataset
     const BASE = '/pages/UI-Pages'
     wx.navigateTo({
-      url: `${BASE}${url}`,
+      url: `${BASE}/${url}/${url}`,
     })
   },