+++ CHANGELOG.md
@@ -1,11 +1,15 @@
 # 版本更新 
 
 ## v0.2
+### 2021/08/10
+- edit code style
+- add `drawer component`
+
 ### 2021/08/03
-- docs: improve picker.md
+- improve picker.md
 
 ### 2021/08/02
-- feat: add dateSelector mode and timeSelector mode to picker component
+- add dateSelector mode and timeSelector mode to picker component
 
 ### 2021/08/01
 - fix css style error of the picker component